feat: initialize application building.
Initialize the application building and inject it to IoC.
All the information is from the task description:
Initially the building should have 2 apartments, apartment 101 and 102, a Gym and a Library.
The building should also have a requested temperature of 25.0 degrees.
Rooms have heating or cooling enabled based on the following logic:
If the room temperature is below the requested building temperature, heating should be enabled.
If the room temperature is above the requested building temperature, cooling should be enabled.